requirements-airflow.txt
# PySpark and Data Processing
pyspark==3.5.0
pyarrow==14.0.1
pandas==2.0.3

# API and Web
requests==2.31.0
redis==5.0.1

# Data Validation
great-expectations==0.17.23

# Visualization
plotly==5.18.0
matplotlib==3.7.5
seaborn==0.13.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
tenacity==8.2.3
tqdm==4.66.1

# Airflow (only for Airflow containers)
apache-airflow-providers-apache-spark==4.5.0
